Module iec::ecs

Super simple ECS-style building blocks, tailored for managing the various side-tables and data structures generated during the compilation process.

Structs

Container
A fancy lookup table mapping for associating a Component with a particular EntityId.

EntityId
An opaque ID used to represent an entity.

Read
A read-only reference to a Container of Components.

ReadWrite
A reference to a Container of Components which supports mutation.

Resources
A resource container used to access the various components stored inside.

Singleton
An immutable reference to a singleton component.

SingletonMut
A mutable reference to a singleton component.

Traits

Component
Abstract component type.

FromResources
An adaptor trait for retrieving a particular Component container from the world.