Trait iec::ecs::FromResources

pub trait FromResources<'r>: Sized {
    fn from_resources(r: &'r Resources) -> Self;

    fn ensure_registered(r: &mut Resources);
}

An adaptor trait for retrieving a particular Component container from the world.

This enables nice things like passing a tuple of Components to a Pass.

Required methods

fn from_resources(r: &'r Resources) -> Self

fn ensure_registered(r: &mut Resources)

Loading content...